updates after rebase on master

+++ schema/deploy/view_ciip_facility.sql
@@ -12,7 +12,7 @@ BEGIN;
     )
     select
        x.id as application_id,
-       x.facility_data ->> 'bcghgid' as bcghgid,
+       (x.facility_data ->> 'bcghgid')::numeric as bcghgid,
        (x.facility_data ->> 'latitude')::numeric as latitude,
        (x.facility_data ->> 'longitude')::numeric as longitude,
        (x.facility_data ->> 'naics_code')::numeric as naics_code,
